server/pyproject.toml
[project]
name = "meddlingkids-server"
version = "0.1.0"
description = "Python server for Meddling Kids"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "agent-framework-core==1.0.0b260130",
    "azure-monitor-opentelemetry-exporter>=1.0.0b35",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.0",
    "aiohttp>=3.13.0",
    "brotli>=1.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.9.1",
    "Pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.35.0",
    "mypy>=1.15.0",
    "ruff>=0.11.0",
]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ── Poe tasks ───────────────────────────────────────────────────

[tool.poe.tasks]

[tool.poe.tasks.lint]
help = "Run all linting (ruff check + mypy)"
sequence = ["lint:ruff", "lint:mypy"]

[tool.poe.tasks."lint:ruff"]
help = "Run ruff linter and formatter check"
sequence = [
    { cmd = "ruff check src/" },
    { cmd = "ruff format --check src/" },
]

[tool.poe.tasks."lint:mypy"]
help = "Run mypy type checking"
cmd = "mypy src/"

[tool.poe.tasks.format]
help = "Auto-fix ruff lint issues and format code"
sequence = [
    { cmd = "ruff check --fix src/" },
    { cmd = "ruff format src/" },
]

# ── Ruff ────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 140
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "T20",   # flake8-print (flag bare prints)
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "T201",   # allow print() — used extensively in logger
    "SIM108", # allow if/else instead of ternary
    "RUF001", # ambiguous unicode in strings (intentional in LLM prompts)
    "RUF002", # ambiguous unicode in docstrings (en dashes, math symbols)
    "RUF003", # ambiguous unicode in comments (intentional)
    "N815",   # mixedCase in class scope (Pydantic models matching JSON API)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ── Mypy ────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
packages = ["src"]
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
